scripts/requirements.txt
# Required Python packages for Archon Prime scripts
python-dotenv
requests
PyYAML
pytest
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
